Backend/parakeet-service/requirements.txt
# Parakeet v3 Speech-to-Text Service Requirements

# FastAPI and web server
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0

# PyTorch (CUDA 12.1 support)
# Install via: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu121
torch>=2.1.0
torchaudio>=2.1.0

# NVIDIA NeMo Toolkit
nemo_toolkit[asr]==2.0.0

# Audio processing
librosa==0.10.1
soundfile==0.12.1
numpy==1.24.3

# Configuration
omegaconf==2.3.0

# Utilities
requests==2.31.0
